apps/battery-analyzer/requirements.txt
aio-pika==9.4.1
annotated-types==0.6.0
APScheduler==3.10.4
black==24.3.0
blinker==1.7.0
fastapi==0.110.1
flake8==7.0.0
motor==3.4.0
numpy==1.26.4
pika==1.3.2
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.3.0
pylint==3.1.0
pymongo==4.6.3
pynvim==0.4.3
pytz==2024.1
virtualenv==20.25.1
urllib3==1.25.8
uvicorn==0.29.0
